feat: only prompt if user not in graphical environment

+++ postinstall.sh
@@ -45,10 +45,12 @@ wal -i "$selected_wallpaper" -n && \
 	source ~/.cache/wal/colors.sh && \
 	cp -r $wallpaper ~/wallpapers/pywallpaper.jpg
 
-echo "jwmarb's Dotfiles successfully installed. Restart your pc to apply changes."
+echo "jwmarb's Dotfiles successfully installed."
 
-read -p "Would you like to restart your pc now? [Y/n]: " choice
+if [[ -z "$DISPLAY" ]]; then
+	read -p "Would you like to restart your pc now? [Y/n]: " choice
 
-if [[ "$choice" =~ ^[Yy]$ ]]; then
-  sudo reboot
+	if [[ "$choice" =~ ^[Yy]$ ]]; then
+	  sudo reboot
+	fi
 fi
